% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_skip}
\alias{calculate_skip}
\title{Calculate lines of data to skip}
\usage{
calculate_skip(
  filename,
  header_rows,
  min_timestamp,
  parse_function = ymd_hms,
  quiet = FALSE
)
}
\arguments{
\item{filename}{Filename, character}

\item{header_rows}{Number of header rows, integer}

\item{min_timestamp}{Minimum timestamp caller is looking for, character}

\item{parse_function}{Parse function to handle timestamps}

\item{quiet}{Print diagnostic message? Logical}
}
\value{
The number of lines OF THE DATA (not of the file) to skip.
If all data should be skipped, a -1 is returned.
}
\description{
This internal function returns the number of lines of a
file to skip based on a requested minimum timestamp to read. It does so
by 'peeking' at the first and last timestamps of the data, ignoring any
header lines. We assume that the timestamps of the
data are in ascending order and regular in their spacing.
}
\note{
For high performance this uses the \code{fpeek} package.
}
