% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_df}
\alias{expand_df}
\title{Expand a data frame by looking for string expansion patterns}
\usage{
expand_df(df)
}
\arguments{
\item{df}{The data frame to expand}
}
\value{
The expanded data frame.
}
\description{
Expand a data frame: look for patterns like x,y,z within
entries and, for that row, replicate it and use expand_string to break
apart the x,y,z. Multiple expansions within a row are OK as long as
they're the same length.
}
