% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand_string}
\alias{expand_string}
\title{Expand a string based on comma or colon patterns}
\usage{
expand_string(s, expand_comma = TRUE, expand_colon = TRUE, quiet = TRUE)
}
\arguments{
\item{s}{String}

\item{expand_comma}{Look for comma expansions ("{X,Y,...}")? Logical}

\item{expand_colon}{Look for colon expansions ("{X:Y}")? Logical}

\item{quiet}{Be quiet or print diagnostic messages? Logical}
}
\value{
The expanded string, as a character vector
}
\description{
look for patterns like "{x,y,z}" within a possibly
}
\examples{
expand_string("{A,B2,C}")
expand_string("B0{1:3}")
}
