% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_directories}
\alias{list_directories}
\title{A recursive function to print a nicely-formatted directory tree and its files}
\usage{
list_directories(
  dir_list,
  outfile = "",
  prefix = "",
  pattern = NULL,
  list_files = TRUE
)
}
\arguments{
\item{dir_list}{A list of directories}

\item{outfile}{Optional file to write output to}

\item{prefix}{Line prefix; used by the function as it recurses}

\item{pattern}{Optional file pattern}

\item{list_files}{List files in each directory? TRUE by default}
}
\value{
Nothing; used only for its side effects (i.e. printing).
}
\description{
A recursive function to print a nicely-formatted directory tree and its files
}
\examples{
\dontrun{
list_directories(list("pipeline/Raw/", "pipeline/L0/",
                      "pipeline/L1_normalize/",
}
}
