% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lx_variables}
\alias{Lx_variables}
\title{COMPASS-FME sensor data variables}
\format{
\subsection{\code{Lx_variables}}{

A data frame with 154 rows and 2 columns:
\describe{
\item{research_name}{Name of the output variable}
\item{final_units}{Units}
\item{description}{Description}
}
}
}
\source{
See https://github.com/COMPASS-DOE/sensor-data-pipeline
}
\usage{
Lx_variables
}
\description{
A listing of the data variables (the "research names")
output by the COMPASS-FME sensor data processing pipeline.
}
\references{
Pennington et al. (2025), "A Performant, Scalable Processing
Pipeline for High-Quality and FAIR Environmental Sensor
Data". https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2025JG008807
}
\keyword{datasets}
