% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.r
\name{nearest_neighbor_TMP}
\alias{nearest_neighbor_TMP}
\title{Find the nearest-neighbor grid squares - TEMPEST}
\usage{
nearest_neighbor_TMP(location, radius = 1)
}
\arguments{
\item{location}{Location grid cell ID (A1 to J8), character}

\item{radius}{Radius desired, 0 or greater (see examples), numeric}
}
\value{
A vector of grid square IDs in a box of radius \code{radius}
}
\description{
Find the nearest-neighbor grid squares - TEMPEST
}
\examples{
nearest_neighbor_TMP("C2")
nearest_neighbor_TMP("A1") # returns four squares (other are out)
nearest_neighbor_TMP("F3", radius = 0) # returns F3
nearest_neighbor_TMP("F3", radius = 1) # returns the nine squares around F3

# Construct a data frame of neighboring grid squares for a vector of locations
locations <- c("C2", "F3")
grid_map <- lapply(locations, function(x) data.frame(neighbor = nearest_neighbor_TMP(x)))
names(grid_map) <- locations
dplyr::bind_rows(grid_map, .id = "location")
}
