% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatroll.R
\name{process_aquatroll_dir}
\alias{process_aquatroll_dir}
\title{Read and process a directory of Aqua TROLL files}
\usage{
process_aquatroll_dir(
  datadir,
  tz,
  dropbox_token = NULL,
  progress_bar = NULL,
  ...
)
}
\arguments{
\item{datadir}{Directory, either in Dropbox or local}

\item{tz}{Time zone the data are set to}

\item{dropbox_token}{Optional Dropbox token}

\item{progress_bar}{Optional progress bar to call while reading}

\item{...}{Other parameters to be passed to \code{\link{read_aquatroll200_file}}
or \code{\link{read_aquatroll600_file}}}
}
\value{
All Aqua TROLL files in directory, read and concatenated, with some
basic processing done: duplicate rows dropped, time zone set, reshaped
to 'long' form, and information such as data logger ID, data logger
channel, and variable number parsed into their separate columns.
}
\description{
Read a directory of Aqua TROLL 200 and/or 600 files,
either from Dropbox or locally.
}
\seealso{
\code{\link{read_aquatroll200_file}} \code{\link{read_aquatroll600_file}}
}
\author{
Ben Bond-Lamberty
}
