% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapflow.R
\name{process_sapflow_dir}
\alias{process_sapflow_dir}
\title{Read and process a directory of sapflow files}
\usage{
process_sapflow_dir(
  datadir,
  tz,
  dropbox_token = NULL,
  progress_bar = NULL,
  ...
)
}
\arguments{
\item{datadir}{Directory, either in Dropbox or local}

\item{tz}{Time zone the data are set to}

\item{dropbox_token}{Optional Dropbox token}

\item{progress_bar}{Optional progress bar to call while reading}

\item{...}{Other parameters to be passed to \code{\link{read_sapflow_file}}}
}
\value{
All sapflow files in directory, read and concatenated, with some
basic processing done: duplicate rows dropped, time zone set, and reshaped
to 'long' form.
}
\description{
Read a directory of sapflow files, either from Dropbox or
locally.
}
\seealso{
\code{\link{read_sapflow_file}}
}
\author{
Ben Bond-Lamberty
}
