% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teros.R
\name{process_teros_dir}
\alias{process_teros_dir}
\title{Read and process a directory of TEROS files}
\usage{
process_teros_dir(datadir, tz, dropbox_token = NULL, progress_bar = NULL, ...)
}
\arguments{
\item{datadir}{Directory, either in Dropbox or local}

\item{tz}{Time zone the data are set to}

\item{dropbox_token}{Optional Dropbox token}

\item{progress_bar}{Optional progress bar to call while reading}

\item{...}{Other parameters to be passed to \code{\link{read_teros_file}}}
}
\value{
All TEROS files in directory, read and concatenated, with some
basic processing done: duplicate rows dropped, time zone set, reshaped
to 'long' form, and information such as data logger ID, data logger
channel, and variable number parsed into their separate columns.
}
\description{
Read a directory of TEROS files, either from Dropbox or
locally.
}
\seealso{
\code{\link{read_teros_file}}
}
\author{
Ben Bond-Lamberty
}
