% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalogger.R
\name{read_datalogger_file}
\alias{read_datalogger_file}
\title{Read a raw Campbell datalogger data file}
\usage{
read_datalogger_file(filename, quiet = FALSE)
}
\arguments{
\item{filename}{Fully-qualified filename of a raw Campbell datalogger file}

\item{quiet}{Print diagnostic messages? Logical}
}
\value{
A \code{\link[tibble]{tibble}} with the data.
}
\description{
This function reads a local file of raw sapflow data,
extracts the logger number from the header, and uses
\code{\link[readr]{read_csv}} to parse the file into a data frame.
}
\examples{
fn <- system.file("PNNL_11_sapflow_1min.dat", package = "compasstools")
read_datalogger_file(fn)
}
\seealso{
\code{\link{process_sapflow_dir}}
}
\author{
Stephanie Pennington and Ben Bond-Lamberty
}
