% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{read_file_dropbox}
\alias{read_file_dropbox}
\title{Download and read a data file from Dropbox}
\usage{
read_file_dropbox(filename, token, read_function)
}
\arguments{
\item{filename}{A Dropbox filename, e.g. returned by \code{drop_dir}}

\item{token}{A dropbox token}

\item{read_function}{A function to read the downloaded file with}
}
\value{
A \code{\link[tibble]{tibble}} with the data.
}
\description{
This function downloads and reads (via a caller-supplied
function) a file on Dropbox.
}
\author{
Ben Bond-Lamberty
}
