% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_L1}
\alias{read_L1}
\title{Read L1 (Level 1) sensor data files}
\usage{
read_L1(variable, site = NULL, path = ".", quiet = FALSE)
}
\arguments{
\item{variable}{Variable name ('research name') to be read, character}

\item{site}{Optional name of the site(s) of data to read, character}

\item{path}{Optional path of the L1 dataset, character}

\item{quiet}{Print diagnostic information? Logical}
}
\value{
A \code{\link[tibble]{tibble}} of L1 data.
}
\description{
This function reads the COMPASS-FME L1 data files (CSV format)
for a single variable, from one or more sites, and returns
the compiled data.
}
\note{
This function only works for L1 v2-0 (July 2025) and higher.
}
\examples{
\dontrun{
read_L1("gw-tds", site = "TMP")
read_L1("gw-tds", c("TMP", "OWC")) # multiple sites
read_L1(variable = "gw-tds") # will read all sites' data
read_L1(variable = "gw-tds", path = "/path/to/L1/data")
}
}
\author{
BBL
}
