% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_L2}
\alias{read_L2}
\title{Read L2 (Level 2) sensor data files (Parquet format)}
\usage{
read_L2(variable, site = NULL, path = ".", quiet = FALSE)
}
\arguments{
\item{variable}{Variable name ('research name') to be read, character}

\item{site}{Optional name of the site(s) of data to read, character}

\item{path}{Optional path of the L2 dataset, character}

\item{quiet}{Print diagnostic information? Logical}
}
\value{
A \code{\link[tibble]{tibble}} of L2 data.
}
\description{
This function reads the COMPASS-FME L2 data files (Parquet format)
for a single variable, from one or more sites, and returns
the compiled data.
}
\note{
This function only works for L2 v2-0 (July 2025) and higher.
}
\examples{
\dontrun{
read_L2("gw-tds", site = "TMP")
read_L2("gw-tds", c("TMP", "OWC")) # multiple sites
read_L2(variable = "gw-tds") # will read all sites' data
read_L2(variable = "gw-tds", path = "/path/to/L2/data")
}
}
\author{
BBL
}
