% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sapflow.R
\name{read_sapflow_dir}
\alias{read_sapflow_dir}
\title{Read a directory of sapflow files, either from Dropbox or locally}
\usage{
read_sapflow_dir(datadir, dropbox_token = NULL, progress_bar = NULL)
}
\arguments{
\item{datadir}{Directory, either in Dropbox or local}

\item{dropbox_token}{Optional Dropbox token}

\item{progress_bar}{Optional progress bar to call while reading}
}
\value{
All sapflow files in directory, read and concatenated.
}
\description{
Read a directory of sapflow files, either from Dropbox or locally
}
\seealso{
read_sapflow_file
}
\author{
Ben Bond-Lamberty
}
