% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sapflow.R
\name{read_sapflow_file}
\alias{read_sapflow_file}
\title{Read a raw sapflow data file}
\usage{
read_sapflow_file(filename)
}
\arguments{
\item{filename}{Fully-qualified filename, character}
}
\value{
A tibble with the data
}
\description{
Read a raw sapflow data file
}
