% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scan_folders}
\alias{scan_folders}
\title{Scan sub-folders}
\usage{
scan_folders(root_dir, file_pattern = "\\\\.csv$", quiet = TRUE)
}
\arguments{
\item{root_dir}{Root directory to start from, character}

\item{file_pattern}{Regex for files, character}

\item{quiet}{Be quiet or print diagnostic messages? Logical}
}
\value{
A named list of folder contents; each list object name is the
name of the folder, and each object in the list is a vector of
fully qualified filenames.
}
\description{
Scan sub-folders
}
\note{
Does not recurse into sub-folders.
}
\examples{
scan_folders("./")
}
