% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lx_variables.R
\name{search_Lx_variables}
\alias{search_Lx_variables}
\title{Search the list of COMPASS-FME sensor variables}
\usage{
search_Lx_variables(term)
}
\arguments{
\item{term}{Search term, case insensitive (character)}
}
\value{
Variables whose descriptions match the search term.
}
\description{
Search the list of COMPASS-FME sensor variables
}
\note{
The search is case-sensitive but definitely not 'smart';
use single words only!
}
\examples{
search_Lx_variables("redox")
search_Lx_variables("this will not find any matches")
}
