% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unit_conversion}
\alias{unit_conversion}
\title{Perform unit conversion based on a data frame of data and units table}
\usage{
unit_conversion(dat, ut, quiet = FALSE)
}
\arguments{
\item{dat}{Data frame with \code{research_name} and \code{value} columns}

\item{ut}{Data frame with \code{research_name}, \code{conversion}, and \code{new_unit} columns}

\item{quiet}{Print progress messages?}
}
\value{
The data frame with new columns \code{value_conv} and \code{units}.
}
\description{
Perform unit conversion based on a data frame of data and units table
}
\examples{
x <- data.frame(value = 1:3, research_name = c("a", "a", "b"))
y <- data.frame(research_name = c("a", "b"), conversion = c("x * 1", "(x * 2) - 1"), new_unit = "")
unit_conversion(x, y)
}
