% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox.R
\name{process_dir}
\alias{process_dir}
\title{Read a directory of files}
\usage{
process_dir(
  datadir,
  pattern,
  read_function,
  dropbox_token = NULL,
  progress_bar = NULL,
  ...
)
}
\arguments{
\item{datadir}{A directory, either in Dropbox or local}

\item{pattern}{Filename regex pattern}

\item{read_function}{The file-read function to use}

\item{dropbox_token}{Optional Dropbox token}

\item{progress_bar}{Optional progress bar to call while reading}

\item{...}{Other parameters to be passed to \code{read_function}}
}
\value{
All data files in directory, read and concatenated
in a \code{\link[tibble]{tibble}}.
}
\description{
A general-purpose function to read data files, either from
Dropbox or locally. The function identifies files based on a pattern,
reads data from each file, and binds them together Callers
pass a \code{read_function} that handles the actual file read.
}
\note{
This is not intended to be called directly by users; use
\code{\link{process_sapflow_dir}}, etc., instead.

We use \code{\link[dplyr]{bind_rows}} for its much greater efficiency
over \code{\link{rbind}}
}
\author{
Ben Bond-Lamberty
}
