% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapflow.R
\name{read_sapflow_file}
\alias{read_sapflow_file}
\title{Read a raw sapflow data file}
\usage{
read_sapflow_file(filename)
}
\arguments{
\item{filename}{Fully-qualified filename, character, of a raw sapflow dataset
from a Campbell datalogger}
}
\value{
A \code{\link[tibble]{tibble}} with the data.
}
\description{
This function reads a local file of raw sapflow data,
extracts the logger number from the header, and uses
\code{\link[readr]{read_csv}} to parse the file into a data frame.
}
\seealso{
\code{\link{process_sapflow_dir}}
}
\author{
Stephanie Pennington
}
