% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapflow.R
\name{read_sapflow_file}
\alias{read_sapflow_file}
\title{Read a raw sapflow data file}
\usage{
read_sapflow_file(filename, min_timestamp = NULL, quiet = FALSE)
}
\arguments{
\item{filename}{Fully-qualified filename of a raw sapflow dataset
from a Campbell datalogger}

\item{min_timestamp}{Minimum timestamp to read, character;
function will skip down in the data until approximately this time}
}
\value{
A \code{\link[tibble]{tibble}} with the data.
}
\description{
This function reads a local file of raw sapflow data,
extracts the logger number from the header, and uses
\code{\link[readr]{read_csv}} to parse the file into a data frame.
}
\examples{
fn <- system.file("PNNL_11_sapflow_1min.dat", package = "compasstools")
read_sapflow_file(fn)
}
\seealso{
\code{\link{process_sapflow_dir}}
}
\author{
Stephanie Pennington
}
