% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teros.R
\name{read_teros_file}
\alias{read_teros_file}
\title{Read a raw TEROS data file}
\usage{
read_teros_file(filename, min_timestamp = NULL)
}
\arguments{
\item{filename}{Fully-qualified filename of a raw TEROS dataset
from a Campbell datalogger}

\item{min_timestamp}{Minimum timestamp to read, character;
function will skip down in the data until approximately this time}
}
\value{
A \code{\link[tibble]{tibble}} with the data.
}
\description{
This function uses
\code{\link[readr]{read_csv}} to parse the file into a data frame.
}
\seealso{
\code{\link{process_teros_dir}}
}
\author{
Stephanie Pennington
}
