% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_sensor_data.R
\name{recent_sensor_data}
\alias{recent_sensor_data}
\title{Get recent sensor data from a COMPASS site.}
\usage{
recent_sensor_data(site, sensor)
}
\arguments{
\item{site}{Site name (CRC, DLG, GWI, OWC, PTR, SWH, or TMP), character}

\item{sensor}{Sensor name, character}
}
\value{
Either a table of available data, if no parameters are specified;
or the requested dataset based on \code{site} and \code{sensor}.
}
\description{
Get recent sensor data from a COMPASS site.
}
\note{
This function is intended for QA/QC and 'quick look' calculations,
not full analyses. The returned data have some metadata attached (e.g.,
\code{Instrument_ID}, \code{research_name}) but are not identical to
what is produced by the sensor data pipeline. In particular, no flagging,
unit conversion, or gap-filling is performed.
}
\examples{
recent_sensor_data() # list available data
recent_sensor_data("DLG", "TEROS12")
}
\author{
BBL
}
